+++ engine/flink/api/src/main/scala/pl/touk/nussknacker/engine/flink/api/NkGlobalParameters.scala
@@ -2,10 +2,14 @@ package pl.touk.nussknacker.engine.flink.api
 
 import java.util
 
+import com.typesafe.config.Config
+
 import scala.collection.JavaConverters._
 import org.apache.flink.api.common.ExecutionConfig
 import org.apache.flink.api.common.ExecutionConfig.GlobalJobParameters
 import pl.touk.nussknacker.engine.api.ProcessVersion
+import net.ceedubs.ficus.readers.ArbitraryTypeReader._
+import net.ceedubs.ficus.Ficus._
 
 //we can use this class to pass config through RuntimeContext to places where it would be difficult to use otherwise
 //Also, those configuration properties will be exposed via Flienk REST API/webconsole
@@ -27,9 +31,19 @@ case class NkGlobalParameters(buildInfo: String,
 }
 
 //this is part of global parameters that is parsed with typesafe Config (e.g. from application.conf/model.conf)
-case class ConfigGlobalParameters(useLegacyMetrics: Option[Boolean])
+case class ConfigGlobalParameters(useLegacyMetrics: Option[Boolean],
+                                  explicitUidInStatefulOperators: Option[Boolean])
 
 object NkGlobalParameters {
+
+  def apply(buildInfo: String, processVersion: ProcessVersion, modelConfig: Config): NkGlobalParameters = {
+    val configGlobalParameters = modelConfig.getAs[ConfigGlobalParameters]("globalParameters")
+    NkGlobalParameters(buildInfo, processVersion, configGlobalParameters)
+  }
+
+  def setInContext(ec: ExecutionConfig, globalParameters: NkGlobalParameters): Unit = {
+    ec.setGlobalJobParameters(globalParameters)
+  }
 
   def readFromContext(ec: ExecutionConfig): Option[NkGlobalParameters] = Option(ec.getGlobalJobParameters).collect {
     case a:NkGlobalParameters => a