+++ engine/flink/util/src/main/scala/pl/touk/nussknacker/engine/flink/util/transformer/PreviousValueTransformer.scala
@@ -5,9 +5,10 @@ import org.apache.flink.api.common.state.{ValueState, ValueStateDescriptor}
 import org.apache.flink.configuration.Configuration
 import org.apache.flink.streaming.api.scala._
 import pl.touk.nussknacker.engine.api._
+import pl.touk.nussknacker.engine.flink.api.compat.ExplicitUidInOperatorsSupport
 import pl.touk.nussknacker.engine.flink.api.process._
 
-object PreviousValueTransformer extends CustomStreamTransformer {
+case object PreviousValueTransformer extends CustomStreamTransformer with ExplicitUidInOperatorsSupport {
 
   type Value = Any
 
@@ -15,10 +16,11 @@ object PreviousValueTransformer extends CustomStreamTransformer {
   def execute(@ParamName("keyBy") keyBy: LazyParameter[String],
               @ParamName("value") value: LazyParameter[Value])
   = FlinkCustomStreamTransformation((start: DataStream[Context], ctx: FlinkCustomNodeContext) =>
-    start
-      .map(ctx.lazyParameterHelper.lazyMapFunction(keyBy))
-      .keyBy(_.value)
-      .map(new PreviousValueFunction(value, ctx.lazyParameterHelper)), value.returnType)
+    setUidToNodeIdIfNeed(ctx,
+      start
+        .map(ctx.lazyParameterHelper.lazyMapFunction(keyBy))
+        .keyBy(_.value)
+        .map(new PreviousValueFunction(value, ctx.lazyParameterHelper))), value.returnType)
 
   class PreviousValueFunction(val parameter: LazyParameter[Value],
                               val lazyParameterHelper: FlinkLazyParameterFunctionHelper) extends RichMapFunction[ValueWithContext[String], ValueWithContext[Any]]